requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-markdown==0.8.4
django-postgrespool==0.3.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
gunicorn==19.3.0
jsonfield==1.0.3
Markdown==2.6.2
static3==0.5.1
whitenoise==2.0.2
